pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-image-recognition"
version = "1.1.0"
description = "MCP server for image recognition using vision APIs (Anthropic, OpenAI, Cloudflare Workers AI)"
readme = "README.md"
authors = [{name = "Mario & Contributors"}]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "mcp>=1.2.0",
    "anthropic>=0.8.0",
    "openai>=1.6.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "pytesseract>=0.3.13",
    "httpx>=0.27.0",
    "requests>=2.30.0",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
mcp-image-recognition = "image_recognition_server.server:main"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 88
target-version = "py310"
